Added build_src to hagrid to choose git branch hagrid uses internally
- removed unused pyproject.toml
