--- packages/grid/backend/pyproject.toml
@@ -1,55 +1,0 @@
-[tool.poetry]
-name = "PyGrid"
-version = "0.7.0-beta.9"
-description = ""
-authors = ["OpenMined <info@openmined.org>"]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-uvicorn = "^0.15.0"
-fastapi = "^0.68.0"
-python-multipart = "^0.0.5"
-email-validator = "^1.0.5"
-requests = "^2.23.0"
-celery = "^5.1.2"
-gevent = "^21.8.0"
-passlib = {extras = ["bcrypt"], version = "^1.7.2"}
-tenacity = "^6.1.0"
-pydantic = "^1.4"
-emails = "^0.5.15"
-raven = "^6.10.0"
-gunicorn = "^20.0.4"
-jinja2 = "^2.11.2"
-psycopg2-binary = "^2.8.5"
-alembic = "^1.4.2"
-sqlalchemy = "^1.3.16"
-pytest = "^5.4.1"
-python-jose = {extras = ["cryptography"], version = "^3.1.0"}
-torch = "1.10.0"
-checksumdir = "^1.2.0"
-websockets = "^9.1.0"
-loguru = "0.5.3"
-
-[tool.poetry.dev-dependencies]
-mypy = "^0.770"
-black = "^19.10b0"
-isort = "^4.3.21"
-autoflake = "^1.3.1"
-flake8 = "^3.7.9"
-pytest = "^5.4.1"
-sqlalchemy-stubs = "^0.3"
-pytest-cov = "^2.8.1"
-pytest-asyncio = "^0.15.1"
-asgi-lifespan = "^1.0.1"
-httpx = "^0.18.2"
-
-[tool.isort]
-multi_line_output = 3
-include_trailing_comma = true
-force_grid_wrap = 0
-line_length = 88
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
-
